Fixing the symbolic link from Scripts/rgis.sh to bin/rgis.

+++ install.sh
@@ -26,5 +26,5 @@ mkdir ${build_dir}
 cd    ${build_dir}
  cmake -DCMAKE_INSTALL_PREFIX="${install_dir}" "${source_dir}"
  make install
- ln -s ${install_dir}/ghaas/Scripts/rgis.sh ${install_dir}/ghaas/bin/rgis
+[ -e  ${install_dir}/ghaas/bin/rgis ] || ln -s ${install_dir}/ghaas/Scripts/rgis.sh ${install_dir}/ghaas/bin/rgis
 rm -rf "${build_dir}"
